+++ test/countries/atlas_engine/lu/validation_transcriber/address_parser_test.rb
@@ -2,11 +2,14 @@
 # frozen_string_literal: true
 
 require "test_helper"
+require "models/atlas_engine/address_validation/address_validation_test_helper"
 
 module AtlasEngine
   module Lu
     module ValidationTranscriber
       class AddressParserTest < ActiveSupport::TestCase
+        include AtlasEngine::AddressValidation::AddressValidationTestHelper
+
         test "One line addresses" do
           [
             [:lu, "4 Op den Aessen", [{ street: "Op den Aessen", building_num: "4" }]],
@@ -17,7 +20,7 @@ module AtlasEngine
             [:lu, "6/8 rue des Bains", [{ street: "rue des Bains", building_num: "6" }]],
             [:lu, "6-8 rue des Bains", [{ street: "rue des Bains", building_num: "6" }]],
           ].each do |country_code, address1, expected|
-            check_parsing(country_code, address1, nil, expected)
+            check_parsing(AddressParser, country_code, address1, nil, expected)
           end
         end
 
@@ -25,26 +28,8 @@ module AtlasEngine
           [
             [:lu, "Maison", "101 Rue Du Cimetiere", [{ street: "Rue Du Cimetiere", building_num: "101" }]],
           ].each do |country_code, address1, address2, expected|
-            check_parsing(country_code, address1, address2, expected)
+            check_parsing(AddressParser, country_code, address1, address2, expected)
           end
-        end
-
-        private
-
-        def check_parsing(country_code, address1, address2, expected, components = nil)
-          components ||= {}
-          components.merge!(country_code: country_code.to_s.upcase, address1: address1, address2: address2)
-          address = AddressValidation::Address.new(**components)
-
-          actual = AddressParser.new(address: address).parse
-
-          assert(
-            expected.to_set.subset?(actual.to_set),
-            "For input ( address1: #{address1.inspect}, address2: #{address2.inspect} )\n\n " \
-              "#{expected.inspect} \n\n" \
-              "Must be included in: \n\n" \
-              "#{actual.inspect}",
-          )
         end
       end
     end